\name{CARE1-package}
\alias{CARE1-package}
\alias{CARE1}
\docType{package}
\title{Statistical package for population size estimation in capture-recapture models.}
\description{The R package CARE1, the first part of the program CARE (Capture-Recapture) in 
http://chao.stat.nthu.edu.tw/softwareCE.html, can be used to analyze epidemiological data via sample coverage approach (Chao et al. 2001a). Based on the input of records from several incomplete lists (or samples) of individuals, the R package CARE1 provides output of population size estimate and related statistics.}
\details{
\tabular{ll}{
Package: \tab CARE1\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2012-10-23\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

}
\author{
T.C. Hsieh\cr
Maintainer: Anne Chao <chao@stat.nthu.edu.tw>
}
\references{
Chao, A., and Tsay, P. K. (1998). A sample coverage approach to multiple-system estimation with application to census undercount. Journal of the American Statistical Association, 93: 283-293. \cr
Chao, A., Tsay, P. K., Lin, S. H., Shau, W. Y., and Chao, D. Y. (2001a). The applications of capture-recapture models to epidemiological data. Statistics in Medicine, 20: 3123-3157. 
}


