\name{Diabetes}
\alias{Diabetes}
\docType{data}
\title{Diabetes data}
\description{Data set on diabetes in a community in Italy.}
\usage{data(Diabetes)}
\format{A four lists data with 15 variables.}
\details{
An epidemiological data with four distinct types of sources, which contains: diabetic clinic and/or family physician visits (1754 cases); 
hospital discharges (452 cases); prescriptions (1135 cases), and purchases of regent strips and insulin syringes (173 cases). 
A total of 2069 cases were identified.
}
\references{
Bruno G, LaPorte R, Merletti F, Biggeri A, McCarty D, Pagano G. 1994. National diabetes programs.
 Application of capture-recapture to count diabetes? Diabetes Care 17(6): 548.
}
\examples{
data(Diabetes)
CARE1.print(Diabetes)
}
\keyword{datasets}
