\name{estN.n}
\alias{estN.n}
\title{Number of identified cases in each records}
\usage{estN.n(z)}
\description{Number of identified cases in each records}
\value{Number of identified cases.}
\arguments{\item{z}{the vector of capture histories or ascertainment records.}}
\examples{data(HAV)
estN.n(HAV)}
