\name{HAV}
\alias{HAV}
\docType{data}
\title{Hepatitis A virus data}
\description{
An outbreak of the hepatitis A virus (HAV) occurred in and around a college in northern Taiwan from April to July 1995.}
\usage{data(HAV)}
\format{A three lists data with 7 variables.}
\details{
Cases of students in that college were ascertained by three sources: 
(i) P-list, records based on a serum test taken by the Institute of Preventive Medicine, Department of Health of Taiwan (135 cases); 
(ii) Q-list, local hospital records reported by the National Quarantine Service (122 cases); 
(iii) E-list, records collected by epidemiologists (126 cases).  Merging the three lists by eliminating duplicate records resulted in 271 ascertained cases.
}

\references{
Chao, D., Shau, W., Lu, C., Chen, K., Chu, C., Shu, H., and Horng, C. (1997). A large outbreak of hepatitis A in a college school in Taiwan: associated with contaminated food and water dissemination. Epidemiology Bulletin, Department of Health, Executive Yuan, Taiwan Government.
}
\examples{data(HAV)}
\keyword{datasets}
